requirements.txt
# Web
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.32.0,<1.0.0
python-multipart>=0.0.12,<1.0.0

# Config & validation
pydantic>=2.9.0,<3.0.0
pydantic-settings>=2.6.0,<3.0.0

# Database
sqlalchemy>=2.0.36,<3.0.0
psycopg2-binary>=2.9.10,<3.0.0
alembic>=1.14.0,<2.0.0

# Cache & messaging
redis>=5.2.0,<6.0.0
aiokafka>=0.11.0,<1.0.0

# RAG / vectors
chromadb>=0.5.20,<1.0.0

# LLM
google-generativeai>=0.8.0,<1.0.0

# Frontend & HTTP client
streamlit>=1.40.0,<2.0.0
httpx>=0.28.0,<1.0.0

# Utilities
python-dotenv>=1.0.1,<2.0.0
structlog>=24.4.0,<26.0.0

# Dev / test
pytest>=8.3.0,<9.0.0
pytest-asyncio>=0.24.0,<1.0.0
pytest-cov>=6.0.0,<7.0.0
